fix useModal
